requirements.txt
Flask>=3.0.0
supabase>=2.4.0
python-dotenv>=1.0.1
opencv-python-headless>=4.13.0
numpy>=2.0.0
Pillow>=10.4.0
